+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `--env` option for declaring env variables in command line. [#8]
+
 ## [0.1.0] - 2021-06-25
 
 ### First Release :tada:
 
 [unreleased]: https://github.com/supertone-inc/envjs/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/supertone-inc/envjs/releases/tag/v0.1.0
+[#8]: https://github.com/supertone-inc/envjs/pull/8